# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Agent Service API
  version: 0.1.0
  description: Versioned public HTTP contract for the workflow and Agent platform.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /workflow/management/process/deploy:
    post:
      tags: [Workflow]
      operationId: deployProcess
      summary: Deploy a BPMN process definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeployProcessRequest' }
      responses:
        '200': { description: Process deployed, content: { application/json: { schema: { $ref: '#/components/schemas/DeployProcessApiResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /workflow/management/process/definition/activate:
    post:
      tags: [Workflow]
      operationId: activateProcessVersion
      summary: Activate a process definition version
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateProcessVersionRequest' }
      responses:
        '200': { description: Version activated, content: { application/json: { schema: { $ref: '#/components/schemas/ActiveProcessVersionApiResponse' } } } }
  /workflow/management/process/definition/active:
    get:
      tags: [Workflow]
      operationId: getActiveProcessVersion
      summary: Read the active process version
      parameters: [{ $ref: '#/components/parameters/ProcessDefinitionKey' }]
      responses:
        '200': { description: Active version, content: { application/json: { schema: { $ref: '#/components/schemas/ActiveProcessVersionApiResponse' } } } }
  /workflow/management/process/definition/exists:
    get:
      tags: [Workflow]
      operationId: processDefinitionExists
      summary: Check whether a process definition exists
      parameters: [{ $ref: '#/components/parameters/ProcessDefinitionKey' }]
      responses:
        '200': { description: Existence result, content: { application/json: { schema: { $ref: '#/components/schemas/BooleanApiResponse' } } } }
  /workflow/management/process/definitions:
    get:
      tags: [Workflow]
      operationId: listProcessDefinitions
      summary: List process definition versions
      parameters:
        - { name: processDefinitionKey, in: query, schema: { type: string } }
      responses:
        '200': { description: Process definitions, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessDefinitionListApiResponse' } } } }
  /workflow/management/process/definition:
    get:
      tags: [Workflow]
      operationId: getProcessDefinition
      summary: Read a process definition
      parameters:
        - { $ref: '#/components/parameters/ProcessDefinitionKey' }
        - { name: version, in: query, schema: { type: integer } }
      responses:
        '200': { description: Process definition, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessDefinitionApiResponse' } } } }
    delete:
      tags: [Workflow]
      operationId: deleteProcessDefinitions
      summary: Delete process definition versions
      parameters: [{ $ref: '#/components/parameters/ProcessDefinitionKey' }]
      responses:
        '200': { description: Definitions deleted, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
  /workflow/management/process/diagram:
    get:
      tags: [Workflow]
      operationId: getProcessInstanceDiagram
      summary: Render a process instance diagram
      parameters: [{ $ref: '#/components/parameters/ProcessInstanceId' }]
      responses:
        '200': { description: PNG diagram, content: { image/png: { schema: { type: string, format: binary } } } }
  /workflow/management/process/diagram-data:
    get:
      tags: [Workflow]
      operationId: getProcessInstanceDiagramData
      summary: Read process instance diagram data
      parameters: [{ $ref: '#/components/parameters/ProcessInstanceId' }]
      responses:
        '200': { description: Diagram data, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessDiagramApiResponse' } } } }
  /workflow/management/process/definitions/page:
    get:
      tags: [Workflow]
      operationId: pageProcessDefinitions
      summary: Page process definitions
      parameters:
        - { $ref: '#/components/parameters/PageNum' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: processDefinitionKey, in: query, schema: { type: string } }
        - { name: processDefinitionName, in: query, schema: { type: string } }
        - { name: publishStatus, in: query, schema: { type: string, default: all } }
      responses:
        '200': { description: Process definition page, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessDefinitionPageApiResponse' } } } }
  /workflow/management/process/instances/page:
    get:
      tags: [Workflow]
      operationId: pageProcessInstances
      summary: Page process instances
      parameters:
        - { $ref: '#/components/parameters/PageNum' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: processDefinitionKey, in: query, schema: { type: string } }
        - { name: processDefinitionName, in: query, schema: { type: string } }
        - { name: processInstanceId, in: query, schema: { type: string } }
        - { name: businessKey, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, default: all } }
      responses:
        '200': { description: Process instance page, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessInstancePageApiResponse' } } } }
  /workflow/management/process/instance/detail:
    get:
      tags: [Workflow]
      operationId: getProcessInstanceDetail
      summary: Read process instance detail
      parameters: [{ $ref: '#/components/parameters/ProcessInstanceId' }]
      responses:
        '200': { description: Process instance detail, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessInstanceDetailApiResponse' } } } }
  /workflow/management/process/instance/terminate:
    post:
      tags: [Workflow]
      operationId: terminateProcessInstance
      summary: Terminate a running process instance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TerminateProcessRequest' }
      responses:
        '200': { description: Process terminated, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
...
components:
  parameters:
    PageNum:
      name: pageNum
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    ProcessDefinitionKey:
      name: processDefinitionKey
      in: query
      required: true
      schema: { type: string }
    ProcessInstanceId:
      name: processInstanceId
      in: query
      required: true
      schema: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: WORKFLOW_ACCESS_TOKEN
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        data: {}
        traceId: { type: string }
    VoidApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: object, nullable: true } }
...
    BooleanApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: boolean } }
...
    DeployProcessApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/DeployProcessResult' } }
    ActiveProcessVersionApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ActiveProcessVersion' } }
    ProcessDefinitionApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessDefinition' } }
    ProcessDefinitionListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: array, items: { $ref: '#/components/schemas/ProcessDefinition' } } }
    ProcessDefinitionPageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessDefinitionPage' } }
    ProcessInstancePageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessInstancePage' } }
    ProcessInstanceDetailApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessInstanceDetail' } }
    ProcessDiagramApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessDiagram' } }
...
    PageMetadata:
      type: object
      required: [total, pageNum, pageSize, records]
      properties:
        total: { type: integer, format: int64 }
        pageNum: { type: integer }
        pageSize: { type: integer }
        records: { type: array, items: {} }
...
    DeployProcessResult:
      type: object
      required: [deploymentId, processDefinitionId, processDefinitionKey, version]
      properties:
        deploymentId: { type: string }
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        processDefinitionName: { type: string, nullable: true }
        version: { type: integer }
    ActiveProcessVersion:
      type: object
      required: [tenantId, processDefinitionKey, processDefinitionId, version, activatedBy, activatedAt]
      properties:
        tenantId: { type: string }
        processDefinitionKey: { type: string }
        processDefinitionId: { type: string }
        version: { type: integer }
        activatedBy: { type: string }
        activatedAt: { type: string, format: date-time }
    ProcessDefinition:
      type: object
      required: [processDefinitionId, processDefinitionKey, version, deploymentId, tenantId, active]
      properties:
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        processDefinitionName: { type: string, nullable: true }
        version: { type: integer }
        deploymentId: { type: string }
        deployedAt: { type: string, format: date-time, nullable: true }
        tenantId: { type: string }
        active: { type: boolean }
        bpmnXml: { type: string, nullable: true }
    ProcessDefinitionSummary:
      type: object
      required: [processDefinitionId, processDefinitionKey, latestVersion, publishStatus, tenantId]
      properties:
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        processDefinitionName: { type: string, nullable: true }
        latestVersion: { type: integer }
        latestDeploymentId: { type: string }
        latestDeployTime: { type: string, format: date-time, nullable: true }
        activeVersion: { type: integer, nullable: true }
        activeProcessDefinitionId: { type: string, nullable: true }
        activeDeploymentId: { type: string, nullable: true }
        activeUpdateTime: { type: string, format: date-time, nullable: true }
        publishStatus: { type: string }
        tenantId: { type: string }
    ProcessDefinitionPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties: { records: { type: array, items: { $ref: '#/components/schemas/ProcessDefinitionSummary' } } }
    ProcessInstanceSummary:
      type: object
      required: [processInstanceId, processDefinitionId, processDefinitionKey, status, tenantId]
      properties:
        processInstanceId: { type: string }
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        processDefinitionName: { type: string, nullable: true }
        processDefinitionVersion: { type: integer, nullable: true }
        latestTaskId: { type: string, nullable: true }
        businessKey: { type: string, nullable: true }
        startUserId: { type: string, nullable: true }
        startTime: { type: string, format: date-time, nullable: true }
        endTime: { type: string, format: date-time, nullable: true }
        lastUpdateTime: { type: string, format: date-time, nullable: true }
        durationInMillis: { type: integer, format: int64, nullable: true }
        deleteReason: { type: string, nullable: true }
        status: { type: string }
        tenantId: { type: string }
    ProcessInstancePage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties: { records: { type: array, items: { $ref: '#/components/schemas/ProcessInstanceSummary' } } }
    ProcessInstanceDetail:
      type: object
      required: [instance, tasks, variables]
      properties:
        instance: { $ref: '#/components/schemas/ProcessInstanceSummary' }
        tasks: { type: array, items: { type: object } }
        variables: { type: array, items: { type: object } }
    ProcessDiagram:
      type: object
      required: [bpmnXml, completedActivityIds, activeActivityIds, highlightedFlows, activityDetails]
      properties:
        bpmnXml: { type: string }
        completedActivityIds: { type: array, items: { type: string } }
        activeActivityIds: { type: array, items: { type: string } }
        highlightedFlows: { type: array, items: { type: string } }
        activityDetails: { type: object, additionalProperties: { type: object } }
...
    DeployProcessRequest:
      type: object
      required: [processDefinitionKey, bpmnXml]
      properties:
        processDefinitionKey: { type: string }
        bpmnXml: { type: string }
        processDefinitionName: { type: string }
    ActivateProcessVersionRequest:
      type: object
      required: [processDefinitionKey, version]
      properties:
        processDefinitionKey: { type: string }
        version: { type: integer }
        processDefinitionId: { type: string }
...
    TerminateProcessRequest:
      type: object
      required: [processInstanceId]
      properties:
        processInstanceId: { type: string }
        reason: { type: string }
...
  responses:
...
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
...